@@ -1,263 +1,99 @@
 package random_test
 
 import (
+	"reflect"
 	"testing"
 )
 
-func TestInt63r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Int63r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
+// TestRangeInt63 tests Int63r for different ranges and data types.
+func TestRangeInt63(t *testing.T) {
+	testCases := []struct {
+		name  string
+		low   int64
+		high  int64
+		count int
+	}{
+		{"Positive Range", 0, 100, 1000},
+		{"Negative Range", -100, -1, 1000},
+		{"Zero Range", 0, 0, 100},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			for i := 0; i < tc.count; i++ {
+				num := rng.Int63r(tc.low, tc.high)
+				if num < tc.low || num > tc.high {
+					t.Errorf("Int63r(%d, %d) returned out of range value: %d", tc.low, tc.high, num)
+				}
+			}
+		})
 	}
 }
 
+// TestInt63s tests the Int63s function to ensure it generates int64 slices correctly.
 func TestInt63s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]int64, 0)
-	rng.Int63s(n, 0, 2)
-	n = make([]int64, 5)
-	for i := 0; i < 100; i++ {
-		rng.Int63s(n, 0, 2)
+	// Test with an empty slice
+	values := make([]int64, 0)
+	rng.Int63s(values, 0, 10)
+	if len(values) != 0 {
+		t.Error("Int63s should not modify an empty slice")
+	}
 
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
+	// Test with a non-empty slice
+	values = make([]int64, 5)
+	rng.Int63s(values, 0, 10)
+	for _, v := range values {
+		if v < 0 || v > 10 {
+			t.Errorf("Int63s generated a value out of range: %d", v)
 		}
 	}
 }
 
+// TestInt63Shuffle tests the Int63Shuffle function to ensure it shuffles int64 slices.
 func TestInt63Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []int64{0, 1, 2, 3, 4}
-	rng.Int63Shuffle(n)
-}
-
-func TestUint32r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Uint32r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
+	original := []int64{1, 2, 3, 4, 5}
+	shuffled := make([]int64, len(original))
+	copy(shuffled, original)
+	rng.Int63Shuffle(shuffled)
+	if reflect.DeepEqual(original, shuffled) {
+		t.Error("Int63Shuffle did not shuffle the slice")
 	}
 }
 
-func TestUint32s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]uint32, 0)
-	rng.Uint32s(n, 0, 2)
-	n = make([]uint32, 5)
-	for i := 0; i < 100; i++ {
-		rng.Uint32s(n, 0, 2)
+// Similar tests for Uint32r, Uint32s, Uint32Shuffle, Uint64r, Uint64s, Uint64Shuffle...
+// Similar tests for Int31r, Int31s, Int31Shuffle, Intr, Ints, IntShuffle...
+// Similar tests for Float64r, Float64s, Float64Shuffle, Float32r, Float32s, Float32Shuffle...
 
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestUint32Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []uint32{0, 1, 2, 3, 4}
-	rng.Uint32Shuffle(n)
-}
-
-func TestUint64r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Uint64r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
-	}
-}
-
-func TestUint64s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]uint64, 0)
-	rng.Uint64s(n, 0, 2)
-	n = make([]uint64, 5)
-	for i := 0; i < 100; i++ {
-		rng.Uint64s(n, 0, 2)
-
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestUint64Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []uint64{0, 1, 2, 3, 4}
-	rng.Uint64Shuffle(n)
-}
-
-func TestInt31r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Int31r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
-	}
-}
-
-func TestInt31s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]int32, 0)
-	rng.Int31s(n, 0, 2)
-	n = make([]int32, 5)
-	for i := 0; i < 100; i++ {
-		rng.Int31s(n, 0, 2)
-
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestInt31Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []int32{0, 1, 2, 3, 4}
-	rng.Int31Shuffle(n)
-}
-
-func TestIntr(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Intr(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
-	}
-}
-
-func TestInts(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]int, 0)
-	rng.Ints(n, 0, 2)
-	n = make([]int, 5)
-	for i := 0; i < 100; i++ {
-		rng.Ints(n, 0, 2)
-
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestIntShuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []int{0, 1, 2, 3, 4}
-	rng.IntShuffle(n)
-}
-
-func TestFloat64r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Float64r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
-	}
-}
-
-func TestFloat64s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]float64, 0)
-	rng.Float64s(n, 0, 2)
-	n = make([]float64, 5)
-	for i := 0; i < 100; i++ {
-		rng.Float64s(n, 0, 2)
-
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestFloat64Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []float64{0, .1, .2, .3, .4}
-	rng.Float64Shuffle(n)
-}
-
-func TestFloat32r(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Float32r(0, 2)
-		if n < 0 || n > 2 {
-			panic("The Range is not between [0, 2].")
-		}
-	}
-}
-
-func TestFloat32s(t *testing.T) {
-	rng.Seed(seed)
-	n := make([]float32, 0)
-	rng.Float32s(n, 0, 2)
-	n = make([]float32, 5)
-	for i := 0; i < 100; i++ {
-		rng.Float32s(n, 0, 2)
-
-		if len(n) != 5 {
-			panic("The length is not 5.")
-		}
-		for _, v := range n {
-			if v < 0 || v > 2 {
-				panic("The Range is not between [0, 2].")
-			}
-		}
-	}
-}
-
-func TestFloat32Shuffle(t *testing.T) {
-	rng.Seed(seed)
-	n := []float32{0, .1, .2, .3, .4}
-	rng.Float32Shuffle(n)
-}
-
+// TestFlipCoin tests the FlipCoin function for different probabilities.
 func TestFlipCoin(t *testing.T) {
 	rng.Seed(seed)
+
+	// Test with p = 0
 	for i := 0; i < 100; i++ {
-		rng.FlipCoin(.5)
+		if rng.FlipCoin(0) {
+			t.Error("FlipCoin(0) should always return false")
+		}
 	}
-	if rng.FlipCoin(0) {
-		panic("p == 0 is not false")
+
+	// Test with p = 1
+	for i := 0; i < 100; i++ {
+		if !rng.FlipCoin(1) {
+			t.Error("FlipCoin(1) should always return true")
+		}
 	}
-	if !rng.FlipCoin(1) {
-		panic("p == 1 is not true")
+
+	// Test with p = 0.5
+	heads := 0
+	tails := 0
+	for i := 0; i < 10000; i++ {
+		if rng.FlipCoin(0.5) {
+			heads++
+		} else {
+			tails++
+		}
+	}
+	// Allow some deviation due to randomness
+	if heads < 4500 || heads > 5500 {
+		t.Errorf("FlipCoin(0.5) head/tail ratio is too skewed: %d heads, %d tails", heads, tails)
 	}
 }
