@@ -17,133 +17,116 @@ var (
 	rng_safe_mt19937 = random.NewthreadSafeRandom(mt19937.New())
 )
 
-func TestInt63(t *testing.T) {
-	rng.Seed(seed)
-	rng.Int63()
-}
+func TestRandom(t *testing.T) {
+	seed := int64(5489)
+	rng := random.New(rand.NewSource(seed))
+	rngMT := random.New(mt19937.New())
 
-func TestUint64(t *testing.T) {
-	rng.Seed(seed)
-	rng.Uint64()
-}
+	testCases := []struct {
+		name     string
+		rng      *random.Random // Random number generator to test
+		function func() interface{}
+		check    func(interface{}) bool // Function to check the result
+	}{
+		{"Int63", rng, func() interface{} { return rng.Int63() }, func(v interface{}) bool { return v.(int64) >= 0 }},
+		{"Uint64", rng, func() interface{} { return rng.Uint64() }, func(v interface{}) bool { return v.(uint64) >= 0 }},
+		{"Uint32", rng, func() interface{} { return rng.Uint32() }, func(v interface{}) bool { return v.(uint32) >= 0 }},
+		{"Int31", rng, func() interface{} { return rng.Int31() }, func(v interface{}) bool { return v.(int32) >= 0 }},
+		{"Int", rng, func() interface{} { return rng.Int() }, func(v interface{}) bool { return v.(int) >= 0 }},
+		{"Float64", rng, func() interface{} { return rng.Float64() }, func(v interface{}) bool { return v.(float64) >= 0 && v.(float64) < 1 }},
+		{"Float32", rng, func() interface{} { return rng.Float32() }, func(v interface{}) bool { return v.(float32) >= 0 && v.(float32) < 1 }},
+		// Add more test cases as needed
+		{"Int63 (MT19937)", rngMT, func() interface{} { return rngMT.Int63() }, func(v interface{}) bool { return v.(int64) >= 0 }},
+		{"Uint64 (MT19937)", rngMT, func() interface{} { return rngMT.Uint64() }, func(v interface{}) bool { return v.(uint64) >= 0 }},
+		{"Uint32 (MT19937)", rngMT, func() interface{} { return rngMT.Uint32() }, func(v interface{}) bool { return v.(uint32) >= 0 }},
+		{"Int31 (MT19937)", rngMT, func() interface{} { return rngMT.Int31() }, func(v interface{}) bool { return v.(int32) >= 0 }},
+		{"Int (MT19937)", rngMT, func() interface{} { return rngMT.Int() }, func(v interface{}) bool { return v.(int) >= 0 }},
+		{"Float64 (MT19937)", rngMT, func() interface{} { return rngMT.Float64() }, func(v interface{}) bool { return v.(float64) >= 0 && v.(float64) < 1 }},
+		{"Float32 (MT19937)", rngMT, func() interface{} { return rngMT.Float32() }, func(v interface{}) bool { return v.(float32) >= 0 && v.(float32) < 1 }},
+		// Add more MT19937 test cases as needed
+	}
 
-func TestUint32(t *testing.T) {
-	rng.Seed(seed)
-	rng.Uint32()
-}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			// Check if the return value meets the expectation
+			if !tc.check(tc.function()) {
+				t.Errorf("%s failed", tc.name)
+			}
+		})
+	}
 
-func TestInt31(t *testing.T) {
-	rng.Seed(seed)
-	rng.Int31()
-}
+	// Test panic cases for Intn family functions
+	panicTestCases := []struct {
+		name     string
+		rng      *random.Random
+		function func()
+	}{
+		{"Int63n", rng, func() { rng.Int63n(-1) }},
+		{"Int31n", rng, func() { rng.Int31n(-1) }},
+		{"Intn", rng, func() { rng.Intn(-1) }},
+		{"Int63n (MT19937)", rngMT, func() { rngMT.Int63n(-1) }},
+		{"Int31n (MT19937)", rngMT, func() { rngMT.Int31n(-1) }},
+		{"Intn (MT19937)", rngMT, func() { rngMT.Intn(-1) }},
+		// Add more panic test cases as needed
+	}
 
-func TestInt(t *testing.T) {
-	rng.Seed(seed)
-	rng.Int()
-}
-
-func TestInt63n(t *testing.T) {
-	rng.Seed(seed)
-	for n := 0; n < 100; n++ {
-		n := rng.Int63n(2)
-		if n < 0 || n >= 2 {
-			panic("The Range is not between [0, 2).")
-		}
+	for _, tc := range panicTestCases {
+		t.Run(tc.name+"_Panic", func(t *testing.T) {
+			defer func() {
+				if r := recover(); r == nil {
+					t.Errorf("%s did not panic as expected", tc.name)
+				}
+			}()
+			tc.function()
+		})
 	}
 }
 
-func TestInt63n_Panic(t *testing.T) {
-	rng.Seed(seed)
-	defer func() {
-		if r := recover(); r == nil {
-			t.Errorf("The code did not panic.")
-		}
-	}()
-
-	rng.Int63n(-32757)
-}
-
-func TestInt31n(t *testing.T) {
-	rng.Seed(seed)
-	for i := 0; i < 100; i++ {
-		n := rng.Int31n(2)
-		if n < 0 || n >= 2 {
-			panic("The Range is not between [0, 2).")
-		}
-	}
-}
-
-func TestInt31n_Panic(t *testing.T) {
-	rng.Seed(seed)
-	defer func() {
-		if r := recover(); r == nil {
-			t.Errorf("The code did not panic.")
-		}
-	}()
-
-	rng.Int63n(-32757)
-}
-
-func TestIntn(t *testing.T) {
-	rng.Seed(seed)
-	for i := 0; i < 100; i++ {
-		n := rng.Intn(2)
-		if n < 0 || n >= 2 {
-			panic("The Range is not between [0, 2).")
-		}
-	}
-}
-
-func TestIntn_Panic(t *testing.T) {
-	rng.Seed(seed)
-	defer func() {
-		if r := recover(); r == nil {
-			t.Errorf("The code did not panic.")
-		}
-	}()
-
-	rng.Intn(-32757)
-}
-
-func TestFloat64(t *testing.T) {
-	rng.Seed(seed)
-	rng.Float64()
-}
-
-func TestFloat32(t *testing.T) {
-	rng.Seed(seed)
-	rng.Float32()
-}
-
-func TestPerm(t *testing.T) {
-	rng.Seed(seed)
+func TestPermAndShuffle(t *testing.T) {
+	// Test Perm
+	rng := random.New(rand.NewSource(seed))
 	for i := 0; i < 100; i++ {
 		n := rng.Perm(5)
-
 		if len(n) != 5 {
-			panic("The length is not 5.")
+			t.Errorf("Perm: expected length 5, got %d", len(n))
 		}
 		for _, v := range n {
 			if v < 0 || v >= 5 {
-				panic("The Range is not between [0, 5).")
+				t.Errorf("Perm: value out of range [0, 5): %d", v)
 			}
 		}
 	}
-}
 
-func TestShuffle(t *testing.T) {
+	// Test Shuffle
 	rng.Seed(seed)
 	n := rng.Perm(5)
-	rng.Shuffle(len(n), func(i, j int) {
-		n[i], n[j] = n[j], n[i]
-	})
+	original := make([]int, len(n))
+	copy(original, n)
+	rng.Shuffle(len(n), func(i, j int) { n[i], n[j] = n[j], n[i] })
+	if isEqual(original, n) {
+		t.Errorf("Shuffle: did not change the order of the slice")
+	}
 }
 
 func TestRead(t *testing.T) {
-	rng.Seed(seed)
+	// Test Read
+	rng := random.New(rand.NewSource(seed))
 	n := make([]byte, 5)
 	_, err := rng.Read(n)
 	if err != nil {
-		panic(err)
+		t.Errorf("Read: unexpected error: %v", err)
 	}
+}
+
+// Helper function to compare two slices for equality
+func isEqual(a, b []int) bool {
+	if len(a) != len(b) {
+		return false
+	}
+	for i, v := range a {
+		if v != b[i] {
+			return false
+		}
+	}
+	return true
 }
