@@ -5,140 +5,111 @@ import (
 	"sync"
 )
 
+var _ rand.Source64 = (*threadSafeRandom)(nil) // Ensures Rand complies with rand.Source64
+
+// threadSafeRandom provides a random number generator safe for concurrent use.
 type threadSafeRandom struct {
-	lk   sync.Mutex
+	lk   sync.RWMutex
 	rand *rand.Rand
 }
 
-// NewthreadSafeRandom -
-// 	returns:
-//	 --	a new thread safe based on Random.
+// NewthreadSafeRandom returns a new threadSafeRandom.
 func NewthreadSafeRandom(src rand.Source) threadSafeRandom {
 	return threadSafeRandom{
 		rand: rand.New(src),
 	}
 }
 
-// Seed -
-// Seed uses the provided seed value to initialize the generator to a deterministic state.
-// Seed should not be called concurrently with any other Rand method.
+// Seed initializes the generator to a deterministic state.
 func (r *threadSafeRandom) Seed(seed int64) {
+	r.lk.Lock()
 	r.rand.Seed(seed)
+	r.lk.Unlock()
 }
 
-// Int63 -
-// 	returns:
-//	 --	an int64 between [-2^63, 2^63-1].
+// Int63 returns a non-negative pseudo-random int64.
 func (r *threadSafeRandom) Int63() int64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Int63()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Uint64 -
-// 	returns:
-//	 --	an Uint64 between [0, 2^64-1].
+// Uint64 returns a non-negative pseudo-random uint64.
 func (r *threadSafeRandom) Uint64() uint64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Uint64()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Uint32 -
-// 	returns:
-//	 --	an Uint32 between [0, 2^32-1].
+// Uint32 returns a non-negative pseudo-random uint32.
 func (r *threadSafeRandom) Uint32() uint32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Uint32()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int31 -
-// 	returns:
-//	 --	an int64 between [-2^31, 2^31-1].
+// Int31 returns a non-negative pseudo-random int32.
 func (r *threadSafeRandom) Int31() int32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Int31()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int -
-// 	returns:
-//	 --	an Int between [-2^31, 2^31-1].
+// Int returns a non-negative pseudo-random int.
 func (r *threadSafeRandom) Int() int {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Int()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int63n -
-//	input:
-//	 n	-- upper limit.
-// 	returns:
-//	 --	an int64 between [0, n).
-// 	panics:
-//	 --	if n<= 0.
+// Int63n returns a non-negative pseudo-random int64 in [0, n).
+// Panics if n <= 0.
 func (r *threadSafeRandom) Int63n(n int64) int64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Int63n(n)
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int31n -
-//	input:
-//	 n	-- upper limit.
-// 	returns:
-//	 --	an int32 between [0, n).
-// 	panics:
-//	 --	if n<= 0.
+// Int31n returns a non-negative pseudo-random int32 in [0, n).
+// Panics if n <= 0.
 func (r *threadSafeRandom) Int31n(n int32) int32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Int31n(n)
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Intn -
-//	input:
-//	 n	-- upper limit.
-// 	returns:
-//	 --	an int between [0, n).
+// Intn returns a non-negative pseudo-random int in [0, n).
 func (r *threadSafeRandom) Intn(n int) int {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Intn(n)
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Float64 -
-// 	returns:
-//	 --	an float64 between [0.0, 1.0).
+// Float64 returns a pseudo-random float64 in [0.0, 1.0).
 func (r *threadSafeRandom) Float64() float64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Float64()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Float32 -
-//	--	returns:
-//	 --	an float32 between [0.0, 1.0).
+// Float32 returns a pseudo-random float32 in [0.0, 1.0).
 func (r *threadSafeRandom) Float32() float32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.rand.Float32()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Perm -
-// 	returns:
-//	 --	as a slice of n ints in the half-open interval [0...n).
+// Perm returns a slice of n ints in [0, n).
 func (r *threadSafeRandom) Perm(n int) []int {
 	r.lk.Lock()
 	val := r.rand.Perm(n)
@@ -146,20 +117,14 @@ func (r *threadSafeRandom) Perm(n int) []int {
 	return val
 }
 
-// Shuffle - Shuffle pseudo-randomizes the order of elements.
-//	input:
-//	 n	-- the number of elements.
-//	 swap	-- swaps the elements with indexes i and j.
-// 	panic: n < 0
+// Shuffle pseudo-randomizes the order of elements.
 func (r *threadSafeRandom) Shuffle(n int, swap func(i, j int)) {
 	r.lk.Lock()
 	r.rand.Shuffle(n, swap)
 	r.lk.Unlock()
 }
 
-// Read - generates len(p) random bytes and writes them into p.
-//	returns:
-//	 --	len(p) and a nil error.
+// Read generates len(p) random bytes and writes them into p.
 func (r *threadSafeRandom) Read(p []byte) (n int, err error) {
 	r.lk.Lock()
 	n, err = r.rand.Read(p)
@@ -167,26 +132,15 @@ func (r *threadSafeRandom) Read(p []byte) (n int, err error) {
 	return n, err
 }
 
-// Int63r - generates pseudo random int64 between low and high.
-//  input:
-//   low  -- lower limit.
-//   high -- upper limit.
-//  returns:
-//	 -- an int64 between [low, high].
+// Int63r returns a pseudo-random int64 in [low, high].
 func (r *threadSafeRandom) Int63r(low, high int64) int64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.Int63()%(high-low+1) + low
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int63s -
-// generates pseudo Randomom integers between low and high.
-//  input:
-//   low    -- lower limit.
-//   high   -- upper limit.
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Int63s fills the values slice with pseudo-random int64 in [low, high].
 func (r *threadSafeRandom) Int63s(values []int64, low, high int64) {
 	r.lk.Lock()
 	if len(values) < 1 {
@@ -198,8 +152,7 @@ func (r *threadSafeRandom) Int63s(values []int64, low, high int64) {
 	r.lk.Unlock()
 }
 
-// Int63Shuffle -
-// shuffles a slice of integers.
+// Int63Shuffle shuffles a slice of int64.
 func (r *threadSafeRandom) Int63Shuffle(values []int64) {
 	r.lk.Lock()
 	var tmp int64
@@ -213,27 +166,15 @@ func (r *threadSafeRandom) Int63Shuffle(values []int64) {
 	r.lk.Unlock()
 }
 
-// Uint32 -
-// generates pseudo Randomom uint32 between low and high.
-//  input:
-//   low  -- lower limit.
-//   high -- upper limit.
-//  returns:
-//	 -- an uint32 between [low, high].
+// Uint32r returns a pseudo-random uint32 in [low, high].
 func (r *threadSafeRandom) Uint32r(low, high uint32) uint32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.Uint32()%(high-low+1) + low
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Uint32s -
-// generates pseudo Randomom integers between low and high.
-//  input:
-//   low    -- lower limit.
-//   high   -- upper limit.
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Uint32s fills the values slice with pseudo-random uint32 in [low, high].
 func (r *threadSafeRandom) Uint32s(values []uint32, low, high uint32) {
 	r.lk.Lock()
 	if len(values) < 1 {
@@ -245,8 +186,7 @@ func (r *threadSafeRandom) Uint32s(values []uint32, low, high uint32) {
 	r.lk.Unlock()
 }
 
-// Uint32Shuffle -
-// shuffles a slice of integers.
+// Uint32Shuffle shuffles a slice of uint32.
 func (r *threadSafeRandom) Uint32Shuffle(values []uint32) {
 	r.lk.Lock()
 	var tmp uint32
@@ -260,27 +200,15 @@ func (r *threadSafeRandom) Uint32Shuffle(values []uint32) {
 	r.lk.Unlock()
 }
 
-// Uint64r -
-// generates pseudo Randomom uint64 between low and high.
-//  input:
-//   low  -- lower limit.
-//   high -- upper limit.
-//  returns:
-//	 -- an uint64 between [low, high].
+// Uint64r returns a pseudo-random uint64 in [low, high].
 func (r *threadSafeRandom) Uint64r(low, high uint64) uint64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.Uint64()%(high-low+1) + low
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Uint64s -
-// generates pseudo Randomom integers between low and high.
-//  input:
-//   low    -- lower limit.
-//   high   -- upper limit.
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Uint64s fills the values slice with pseudo-random uint64 in [low, high].
 func (r *threadSafeRandom) Uint64s(values []uint64, low, high uint64) {
 	r.lk.Lock()
 	if len(values) < 1 {
@@ -292,8 +220,7 @@ func (r *threadSafeRandom) Uint64s(values []uint64, low, high uint64) {
 	r.lk.Unlock()
 }
 
-// Uint64Shuffle -
-// shuffles a slice of integers.
+// Uint64Shuffle shuffles a slice of uint64.
 func (r *threadSafeRandom) Uint64Shuffle(values []uint64) {
 	r.lk.Lock()
 	var tmp uint64
@@ -307,27 +234,15 @@ func (r *threadSafeRandom) Uint64Shuffle(values []uint64) {
 	r.lk.Unlock()
 }
 
-// Int31r -
-// is int range generates pseudo Randomom int32 between low and high.
-//  input:
-//   low  -- lower limit.
-//   high -- upper limit.
-//  returns:
-//	 -- an int32 between [low, high].
+// Int31r returns a pseudo-random int32 in [low, high].
 func (r *threadSafeRandom) Int31r(low, high int32) int32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.Int31()%(high-low+1) + low
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Int31s -
-// generates pseudo Randomom integers between low and high.
-//  input:
-//   low    -- lower limit.
-//   high   -- upper limit.
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Int31s fills the values slice with pseudo-random int32 in [low, high].
 func (r *threadSafeRandom) Int31s(values []int32, low, high int32) {
 	r.lk.Lock()
 	if len(values) < 1 {
@@ -339,8 +254,7 @@ func (r *threadSafeRandom) Int31s(values []int32, low, high int32) {
 	r.lk.Unlock()
 }
 
-// Int31Shuffle -
-// shuffles a slice of integers.
+// Int31Shuffle shuffles a slice of int32.
 func (r *threadSafeRandom) Int31Shuffle(values []int32) {
 	r.lk.Lock()
 	var tmp int32
@@ -354,27 +268,15 @@ func (r *threadSafeRandom) Int31Shuffle(values []int32) {
 	r.lk.Unlock()
 }
 
-// Intr -
-// is int range generates pseudo Randomom integer between low and high.
-//  input:
-//   low  -- lower limit.
-//   high -- upper limit.
-//  returns:
-//	 -- an int between [low, high].
+// Intr returns a pseudo-random int in [low, high].
 func (r *threadSafeRandom) Intr(low, high int) int {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := r.Int()%(high-low+1) + low
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Ints -
-// generates pseudo Randomom integers between low and high.
-//  input:
-//   low    -- lower limit.
-//   high   -- upper limit.
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Ints fills the values slice with pseudo-random int in [low, high].
 func (r *threadSafeRandom) Ints(values []int, low, high int) {
 	r.lk.Lock()
 	if len(values) < 1 {
@@ -386,8 +288,7 @@ func (r *threadSafeRandom) Ints(values []int, low, high int) {
 	r.lk.Unlock()
 }
 
-// IntShuffle -
-// shuffles a slice of integers.
+// IntShuffle shuffles a slice of int.
 func (r *threadSafeRandom) IntShuffle(values []int) {
 	r.lk.Lock()
 	var j, tmp int
@@ -400,27 +301,15 @@ func (r *threadSafeRandom) IntShuffle(values []int) {
 	r.lk.Unlock()
 }
 
-// Float64r -
-// generates a pseudo Randomom real number between low and high; i.e. in [low, right)
-//  input:
-//   low  -- lower limit. (closed)
-//   high -- upper limit. (open)
-//  returns:
-//	 -- an int between [low, high].
+// Float64r returns a pseudo-random float64 in [low, high).
 func (r *threadSafeRandom) Float64r(low, high float64) float64 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := low + (high-low)*r.Float64()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Float64s -
-// generates pseudo Randomom real numbers between low and high; i.e. in [low, right)
-//  input:
-//   low  -- lower limit. (closed)
-//   high -- upper limit. (open)
-//  output:
-//   values -- slice to be filled with len(values) numbers
+// Float64s fills the values slice with pseudo-random float64 in [low, high).
 func (r *threadSafeRandom) Float64s(values []float64, low, high float64) {
 	r.lk.Lock()
 	for i := 0; i < len(values); i++ {
@@ -429,8 +318,7 @@ func (r *threadSafeRandom) Float64s(values []float64, low, high float64) {
 	r.lk.Unlock()
 }
 
-// Float64Shuffle -
-// shuffles a slice of float point numbers
+// Float64Shuffle shuffles a slice of float64.
 func (r *threadSafeRandom) Float64Shuffle(values []float64) {
 	r.lk.Lock()
 	var tmp float64
@@ -444,26 +332,15 @@ func (r *threadSafeRandom) Float64Shuffle(values []float64) {
 	r.lk.Unlock()
 }
 
-// Float32r generates a pseudo Randomom real number between low and high; i.e. in [low, right)
-//  Input:
-//   low  -- lower limit. (closed)
-//   high -- upper limit. (open)
-//  returns:
-//	 -- an int between [low, high].
+// Float32r returns a pseudo-random float32 in [low, high).
 func (r *threadSafeRandom) Float32r(low, high float32) float32 {
-	r.lk.Lock()
+	r.lk.RLock()
 	val := low + (high-low)*r.Float32()
-	r.lk.Unlock()
+	r.lk.RUnlock()
 	return val
 }
 
-// Float32s -
-// generates pseudo Randomom real numbers between low and high; i.e. in [low, right)
-//  input:
-//   low  -- lower limit. (closed)
-//   high -- upper limit. (open)
-//  output:
-//   values -- slice to be filled with len(values) numbers.
+// Float32s fills the values slice with pseudo-random float32 in [low, high).
 func (r *threadSafeRandom) Float32s(values []float32, low, high float32) {
 	r.lk.Lock()
 	for i := 0; i < len(values); i++ {
@@ -472,8 +349,7 @@ func (r *threadSafeRandom) Float32s(values []float32, low, high float32) {
 	r.lk.Unlock()
 }
 
-// Float32Shuffle -
-// shuffles a slice of float point numbers.
+// Float32Shuffle shuffles a slice of float32.
 func (r *threadSafeRandom) Float32Shuffle(values []float32) {
 	r.lk.Lock()
 	var tmp float32
@@ -487,31 +363,22 @@ func (r *threadSafeRandom) Float32Shuffle(values []float32) {
 	r.lk.Unlock()
 }
 
-// FlipCoin -
-// generates a Bernoulli variable; throw a coin with probability p.
+// FlipCoin returns true with probability p.
 func (r *threadSafeRandom) FlipCoin(p float64) bool {
-	r.lk.Lock()
+	r.lk.RLock()
+	defer r.lk.RUnlock()
+
 	if p == 1.0 {
 		return true
 	}
 	if p == 0.0 {
 		return false
 	}
-	if r.Float64() <= p {
-		return true
-	}
-
-	r.lk.Unlock()
-	return false
+	return r.Float64() <= p
 }
 
-// Uint32n -
-//  input:
-//   n	-- upper limit.
-//  returns:
-//	 -- an uint32 between [0, n).
-// 	panics:
-//	 --	if n<= 0.
+// Uint32n returns a non-negative pseudo-random uint32 in [0, n).
+// Panics if n <= 0.
 func (r *threadSafeRandom) Uint32n(n uint32) uint32 {
 	if n <= 0 {
 		panic("invalid argument to Uint32n")
@@ -519,280 +386,69 @@ func (r *threadSafeRandom) Uint32n(n uint32) uint32 {
 	return r.Uint32() % n
 }
 
-// Float64n -
-//  input:
-//   n	-- upper limit.
-//  returns:
-//	 -- an float64 between [0, n).
+// Float64n returns a pseudo-random float64 in [0.0, n).
 func (r *threadSafeRandom) Float64n(n float64) float64 {
 	return n * r.Float64()
 }
 
-// Float32n -
-//  input:
-//   n	-- upper limit.
-//  returns:
-//	 -- an float32 between [0, n).
-// 	panics:
-//	 --	if n<= 0.
+// Float32n returns a pseudo-random float32 in [0.0, n).
+// Panics if n <= 0.
 func (r *threadSafeRandom) Float32n(n float32) float32 {
-	return (n) * r.Float32()
-}
-
-// Float64w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive float64 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-//   -- if w[i] is not positive.
-func (r *threadSafeRandom) Float64w(w []float64) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Float64w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts float64
-	for _, v := range w {
-		if v < 0 {
-			panic("Float64w: w[i] is not positive.")
-		}
-		totalWeigehts += v
-	}
-
-	dice := r.Float64n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Float32w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive float32 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-//   -- if w[i] is not positive.
-func (r *threadSafeRandom) Float32w(w []float32) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Float32w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts float32
-	for _, v := range w {
-		if v < 0 {
-			panic("Float32w: w[i] is not positive.")
-		}
-		totalWeigehts += v
-	}
-
-	dice := r.Float32n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Uint64w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive uint64 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-func (r *threadSafeRandom) Uint64w(w []uint64) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Uint64w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts uint64
-	for _, v := range w {
-		totalWeigehts += v
-	}
-
-	dice := r.Uint64n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Uint32w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive uint32 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-func (r *threadSafeRandom) Uint32w(w []uint32) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Uint32w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts uint32
-	for _, v := range w {
-		totalWeigehts += v
-	}
-
-	dice := r.Uint32n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Int64w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive int64 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-//   -- if w[i] is not positive.
-func (r *threadSafeRandom) Int64w(w []int64) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Int64w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts int64
-	for _, v := range w {
-		if v < 0 {
-			panic("Int64w: w[i] is not positive.")
-		}
-		totalWeigehts += v
-	}
-
-	dice := r.Int63n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Int32w -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive int32 w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-//   -- if w[i] is not positive.
-func (r *threadSafeRandom) Int32w(w []int32) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Int32w")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts int32
-	for _, v := range w {
-		if v < 0 {
-			panic("Int32w: w[i] is not positive.")
-		}
-		totalWeigehts += v
-	}
-
-	dice := r.Int31n(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Intw -
-// Ｗhich randomly picks an index in the range [0, w.length - 1] (inclusive) and returns it.
-// The probability of picking an index i is w[i] / sum(w).
-//  input:
-//   w	-- an array of positive int w.
-//  returns:
-//	 -- picks an index in the range [0, len(w)-1]..
-// 	panics:
-//	 --	if n<= 0.
-//   -- if w[i] is not positive.
-func (r *threadSafeRandom) Intw(w []int) int {
-	l := len(w)
-	if l == 0 {
-		panic("invalid argument to Intw")
-	}
-	if l == 1 {
-		return 0
-	}
-	var totalWeigehts int
-	for _, v := range w {
-		if v < 0 {
-			panic("Intw: w[i] is not positive.")
-		}
-		totalWeigehts += v
-	}
-
-	dice := r.Intn(totalWeigehts)
-	for i := 0; i < l-1; i++ {
-		if dice < w[i] {
-			return i
-		}
-		dice -= w[i]
-	}
-	return l - 1
-}
-
-// Uint64n -
-//  input:
-//   n	-- upper limit.
-//  returns:
-//	 -- an uint64 between [0, n).
-// 	panics:
-//	 --	if n<= 0.
-func (r *threadSafeRandom) Uint64n(n uint64) uint64 {
 	if n <= 0 {
-		panic("invalid argument to Uint64n")
+		panic("invalid argument to Float32n")
 	}
-	return r.Uint64() % n
+	return n * r.Float32()
+}
+
+// ----------------------------------------------------------------------------
+// Weighted Random Selection
+// ----------------------------------------------------------------------------
+
+// Float64w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty or contains non-positive values.
+func (r *threadSafeRandom) Float64w(w []float64) int {
+	return weightedRandomIndexFloat64(r.rand, w)
+}
+
+// Float32w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty or contains non-positive values.
+func (r *threadSafeRandom) Float32w(w []float32) int {
+	return weightedRandomIndexFloat32(r.rand, w)
+}
+
+// Uint64w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty.
+func (r *threadSafeRandom) Uint64w(w []uint64) int {
+	return weightedRandomIndexUint64(r.rand, w)
+}
+
+// Uint32w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty.
+func (r *threadSafeRandom) Uint32w(w []uint32) int {
+	return weightedRandomIndexUint32(r.rand, w)
+}
+
+// Int64w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty or contains non-positive values.
+func (r *threadSafeRandom) Int64w(w []int64) int {
+	return weightedRandomIndexInt64(r.rand, w)
+}
+
+// Int32w randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty or contains non-positive values.
+func (r *threadSafeRandom) Int32w(w []int32) int {
+	return weightedRandomIndexInt32(r.rand, w)
+}
+
+// Intw randomly picks an index in [0, len(w)-1] based on the weights in slice w.
+// The probability of picking index i is w[i] / sum(w).
+// Panics if w is empty or contains non-positive values.
+func (r *threadSafeRandom) Intw(w []int) int {
+	return weightedRandomIndexInt(r.rand, w)
 }
